feat: db migration, seeding from Render
Modified the necessary configuration files to connect to the database
hosted on Render. Ensure that the migration and seeding script run fine
and the back-end server starts without any issues locally. Also refactor
the .env.example (and the .env) file to include the necessary
environment variables for connection to the database, and organize in a
way that follows best practices. Finally, update the package.json file
to have seprate scripts for when running development and production
environments.