prepare server.js for back-end deployment
Update .env.example and .env for back-end deployment, specifically,
remove the SERVER_HOST from the .env.example and .env files. Also,
remove the specific numbers hard-coded in .env.example, as this can and
be changed by anyone pulling the repo. Finally, update the package.json
to have the 'start' run the 'build' script and start up the server.